+++ bcbio/chipseq/peaks.py
@@ -2,6 +2,7 @@
 """
 import os
 import copy
+import toolz as tz
 
 from bcbio.log import logger
 from bcbio import bam, utils
@@ -46,21 +47,24 @@ def calling(data):
     caller_fn = get_callers()[data["peak_fn"]]
     name = dd.get_sample_name(data)
     out_dir = utils.safe_makedir(os.path.join(dd.get_work_dir(data), data["peak_fn"], name ))
-    # chip_bam = _prepare_bam(chip_bam, dd.get_variant_regions(data), data['config'])
-    # input_bam = _prepare_bam(input_bam, dd.get_variant_regions(data), data['config'])
-    out_file = caller_fn(name, chip_bam, input_bam, dd.get_genome_build(data), out_dir,
+    encode_bed = tz.get_in(["genome_resources", "variation", "encode_blacklist"], data)
+    # lcr_bed = utils.get_in(data, ("genome_resources", "variation", "lcr"))
+    chip_bam = _prepare_bam(chip_bam, encode_bed, data['config'])
+    input_bam = _prepare_bam(input_bam, encode_bed, data['config'])
+    out_files = caller_fn(name, chip_bam, input_bam, dd.get_genome_build(data), out_dir,
                          dd.get_chip_method(data), data["config"])
-    data["peaks_file"] = out_file
+    data.update({"peaks_files": out_files})
     return [[data]]
 
 def _prepare_bam(bam_file, bed_file, config):
+    """Remove regions from bed files"""
     if not bam_file or not bed_file:
         return bam_file
     out_file = utils.append_stem(bam_file, '_filter')
-    samtools = config_utils.get_program("samtools", config)
+    bedtools = config_utils.get_program("bedtools", config)
     if not utils.file_exists(out_file):
         with file_transaction(out_file) as tx_out:
-            cmd = "{samtools} view -bh -L {bed_file} {bam_file} > {tx_out}"
+            cmd = "{bedtools} subtract -nonamecheck -A -a {bam_file} -b {bed_file} > {tx_out}"
             do.run(cmd.format(**locals()), "Clean %s" % bam_file)
     return out_file
 
@@ -71,11 +75,11 @@ def _sync(original, processed):
     to consider multiple callers.
     """
     for original_sample in original:
-        original_sample[0]["peaks_file"] = []
+        original_sample[0]["peaks_files"] = {}
         for processs_sample in processed:
             if dd.get_sample_name(original_sample[0]) == dd.get_sample_name(processs_sample[0]):
-                if utils.file_exists(processs_sample[0]["peaks_file"]):
-                    original_sample[0]["peaks_file"].append(processs_sample[0]["peaks_file"])
+                if processs_sample[0]["peaks_files"]:
+                    original_sample[0]["peaks_files"].update(processs_sample[0]["peaks_files"])
     return original
 
 def _check(sample, data):