+++ bcbio/cwl/create.py
@@ -73,6 +73,19 @@ def _write_tool(step_dir, name, inputs, outputs, parallel, programs, file_estima
            "arguments": [],
            "inputs": [],
            "outputs": []}
+    if programs:
+        def resolve_package(p):
+            out = {}
+            parts = p.split("=")
+            if len(parts) == 2:
+                out["package"] = parts[0]
+                out["version"] = [parts[1]]
+            else:
+                out["package"] = p
+            out["specs"] = ["https://anaconda.org/bioconda/%s" % out["package"]]
+            return out
+        out["hints"].append({"class": "SoftwareRequirement",
+                             "packages": [resolve_package(p) for p in programs]})
     out["arguments"].append({"position": 0, "valueFrom":
                              "sentinel-runtime=cores,$(runtime['cores']),ram,$(runtime['ram'])"})
     std_inputs = [{"id": "sentinel-parallel", "type": "string",
@@ -151,13 +164,15 @@ def _step_template(name, run_file, inputs, outputs, parallel):
                 step_inp[attr] = inp[attr]
         sinputs.append(step_inp)
         # scatter on inputs from previous processes that have been arrayed
-        if parallel in "multi-parallel" or len(inp["id"].split("/")) > 1:
+        if (parallel in "multi-parallel" or len(inp["id"].split("/")) > 1
+              or len(inp.get("source", "").split("/")) > 1):
             scatter_inputs.append(step_inp["id"])
     out = {"run": run_file,
            "id": name,
            "in": sinputs,
            "out": [{"id": workflow.get_base_id(output["id"])} for output in outputs]}
     if parallel in ["single-parallel", "multi-parallel", "batch-parallel"]:
+        assert scatter_inputs, "Did not find items to scatter on: %s" % name
         out.update({"scatterMethod": "dotproduct",
                     "scatter": scatter_inputs})
     return out