+++ bcbio/variation/damage.py
@@ -75,7 +75,7 @@ def _rec_filter_to_info(line):
     for f in parts[6].split(";"):
         if f in move_filters:
             bias_info.append(move_filters[f])
-        elif f not in ".":
+        elif f not in ["."]:
             new_filters.append(f)
     if bias_info:
         parts[7] += ";DKFZBias=%s" % ",".join(bias_info)