+++ HISTORY.md
@@ -1,3 +1,24 @@
+## 1.0.3 (in progress)
+
+- Fix FreeBayes somatic and multi-sample calling order to be consistent between
+  chromosome region runs. Thanks to Ho Danliang.
+- Fix structural variant output upload for complex batching cases. Correctly
+  handle shared normals and other multi-batch by naming outputs using batches.
+  Thanks to Sven-Eric Schelhorn.
+- Several updates to the VarScan support: honor options specified in the
+  resource config section; honor min_allele_frac option and set --strand-filter
+  flag in the single-sample case; general cleanups. Thanks to Christian Brueffer.
+- Update validation plots to support matplotlib 2.0.
+- Enable mixed list/string inputs to germline calling. Thanks to Luca Beltrame.
+- Fix qsignature outfile parsing. Thanks to Oliver Hofmann.
+- Allow specification of seq2c options through resources. Thanks to Sally Luke
+  and Marisa Cunha.
+- Provide absolute paths for relative paths to files in algorithm list inputs.
+  Thanks to Matthias De Smet.
+- Switch to Salmon from Sailfish as default alignment-free RNA-seq
+  quantification algorithm.
+- Add `sailfish` as a valid option for `expression_caller`.
+
 ## 1.0.2 (7 March 2017)
 
 - Fix FreeBayes paired somatic calling by generalizing support for finding